Cambios finales de la entrega

+++ JS/productos.js
@@ -161,6 +161,9 @@ const productos = [
     }
 ]
 
+
+
+
 const productosContainer = document.getElementById("productos-container");
 const searchInput = document.querySelector('.form-control');
 const carritoCantidadElement = document.getElementById('carrito-cantidad');
@@ -169,37 +172,67 @@ const totalCarrito = document.getElementById('total');
 const vaciarCarritoBtn = document.getElementById('vaciar-carrito');
 let carritoCantidad = 0;
 
+function cargarProductos() {
+    return new Promise((resolve, reject) => {
+        fetch("../JS/productos.json")
+            .then((response) => {
+                if (!response.ok) {
+                    throw new Error(`Error al cargar los productos. Código de estado: ${response.status}`);
+                }
+                return response.json();
+            })
+            .then((data) => resolve(data)) 
+            .catch((error) => {
+                console.error("Error al cargar productos:", error);
+                reject(error);
+            });
+    });
+}
 
-// Función para obtener el carrito desde localStorage
 function obtenerCarritoDesdeLocalStorage() {
     const carritoGuardado = localStorage.getItem('carrito');
     return carritoGuardado ? JSON.parse(carritoGuardado) : [];
 }
 
-// Función para guardar el carrito en localStorage
 function guardarCarritoEnLocalStorage(carrito) {
     localStorage.setItem('carrito', JSON.stringify(carrito));
 }
 
-// Función para cargar el carrito desde localStorage al cargar la página
-function cargarCarritoDesdeLocalStorage() {
-    const carritoActual = obtenerCarritoDesdeLocalStorage();
-    carritoActual.forEach(producto => agregarProductoAlCarritoLista(producto));
-    calcularTotal(); // Asegúrate de calcular el total después de cargar el carrito
+async function cargarCarritoDesdeLocalStorage() {
+    try {
+        const productos = await cargarProductos();
+        const carritoActual = obtenerCarritoDesdeLocalStorage();
+        if (carritoActual.length > 0 && productos.length > 0) {
+            listaCarrito.innerHTML = '';
+
+            carritoActual.forEach(producto => {
+                const productoEnLista = productos.find(p => JSON.stringify(p) === JSON.stringify(producto));
+
+                if (productoEnLista) {
+                    agregarProductoAlCarritoLista(producto);
+                }
+            });
+
+            calcularTotal();
+        }
+    } catch (error) {
+        console.error('Error al cargar el carrito desde localStorage:', error);
+    }
 }
 
-function mostrarProductos(categoria, terminoBusqueda) {
-    productosContainer.innerHTML = ""; // Limpiar el contenedor
 
-    // Obtener productos según la categoría seleccionada
-    const productosFiltrados = filtrarProductosPorCategoria(productos, categoria);
 
-    // Filtrar productos por término de búsqueda
-    const productosBuscados = buscarProductos(productosFiltrados, terminoBusqueda || '');
+async function mostrarProductos(categoria, terminoBusqueda) {
+    try {
+        const productos = await cargarProductos();
 
-    // Iterar sobre los productos filtrados y agregar al contenedor
-    productosBuscados.forEach(producto => {
-        const card = `
+        productosContainer.innerHTML = ''; 
+
+        const productosFiltrados = filtrarProductosPorCategoria(productos, categoria);
+        const productosBuscados = buscarProductos(productosFiltrados, terminoBusqueda || '');
+
+        productosBuscados.forEach(producto => {
+            const card = `
             <div class="col-md-4 mb-4">
                 <div class="card h-100">
                     <img src="${producto.imagen}" class="card-img-top" alt="${producto.titulo}">
@@ -207,37 +240,51 @@ function mostrarProductos(categoria, terminoBusqueda) {
                         <h5 class="card-title">${producto.titulo}</h5>
                         <p class="card-text">${producto.categoria.nombre}</p>
                         <p class="card-text">$${producto.precio}</p>
-                        <button class="btn btn-primary" onclick="agregarProductoAlCarrito('${producto.id}', '${producto.titulo}', ${producto.precio})">Agregar al carrito</button>
+                        <button class="btn btn-primary" onclick="agregarProductoAlCarrito('${producto.id}', '${producto.titulo}', ${producto.precio}, '${producto.imagen}')">Agregar al carrito</button>
                     </div>
                 </div>
             </div>
-        `;
+            `;
 
-        productosContainer.innerHTML += card;
-    });
-}
-
-function filtrarProductosPorCategoria(productos, categoria) {
-    if (categoria === 'urbanas') {
-        return productos.filter(producto => producto.categoria.id === 'urbanas');
-    } else if (categoria === 'oversize') {
-        return productos.filter(producto => producto.categoria.id === 'oversizes');
-    } else if (categoria === 'musculosas') {
-        return productos.filter(producto => producto.categoria.id === 'musculosas');
-    } else {
-        return productos;
+            productosContainer.innerHTML += card;
+        });
+    } catch (error) {
+        console.error("Error al cargar y mostrar productos:", error);
     }
 }
 
-// Función para buscar productos por término
+function filtrarProductosPorCategoria(productos, categoria) {
+    return categoria ? productos.filter(producto => producto.categoria.id === categoria) : productos;
+}
+
 function buscarProductos(productos, termino) {
     return productos.filter(producto =>
         producto.titulo && producto.titulo.toLowerCase().includes(termino.toLowerCase())
     );
 }
 
+function mostrarMensajeAgregado() {
+    const mensaje = document.createElement('div');
+    mensaje.textContent = '¡Tu producto fue agregado!';
+    mensaje.classList.add('mensaje-agregado');
+    document.body.appendChild(mensaje);
 
-// Función para agregar un producto al carrito
+    mensaje.style.position = 'fixed';
+    mensaje.style.top = '10px';
+    mensaje.style.left = '50%';
+    mensaje.style.transform = 'translateX(-50%)';
+    mensaje.style.padding = '10px';
+    mensaje.style.backgroundColor = 'lightblue';
+    mensaje.style.border = '1px solid #ccc';
+    mensaje.style.borderRadius = '5px';
+
+    setTimeout(() => {
+        mensaje.remove();
+    }, 1500);
+}
+
+
+
 function agregarProductoAlCarrito(idProducto, nombreProducto, precioProducto, imagenProducto) {
     carritoCantidad++;
     actualizarIconoCarrito();
@@ -246,27 +293,64 @@ function agregarProductoAlCarrito(idProducto, nombreProducto, precioProducto, im
     carritoActual.push(nuevoProducto);
     guardarCarritoEnLocalStorage(carritoActual);
     agregarProductoAlCarritoLista(nuevoProducto);
+    mostrarMensajeAgregado();
 }
 
-// Función para agregar un producto al listado del carrito
+
 function agregarProductoAlCarritoLista(producto) {
-    const elementoCarrito = document.createElement('li');
-    elementoCarrito.innerHTML = `
-        <div class="carrito-item">
-            <img src="${producto.imagen}" alt="${producto.nombre}" class="carrito-imagen">
-            <div class="carrito-info">
-                <p>${producto.nombre} - $${producto.precio.toFixed(2)}</p>
-                <button class="eliminar-producto" data-id="${producto.id}" onclick="eliminarProducto(event)">Eliminar</button>
-            </div>
-        </div>
-    `;
+    const productosEnCarrito = listaCarrito.querySelectorAll('.carrito-item');
+    let productoExistente = null;
 
-listaCarrito.appendChild(elementoCarrito);
-calcularTotal();
+    productosEnCarrito.forEach((item) => {
+        const idExistente = item.querySelector('.eliminar-producto').getAttribute('data-id');
+        if (idExistente === producto.id) {
+            productoExistente = item;
+        }
+    });
+
+    if (productoExistente) {
+        // Actualizar la cantidad 
+        const cantidadElemento = productoExistente.querySelector('.cantidad-producto');
+        const cantidadActual = parseInt(cantidadElemento.textContent, 10);
+        cantidadElemento.innerHTML = `<strong>${cantidadActual + 1}</strong>`;
+
+        const precioElemento = productoExistente.querySelector('.precio-total');
+        const precioActual = parseFloat(precioElemento.textContent);
+        precioElemento.textContent = (precioActual + producto.precio).toFixed(2);
+    } else {
+        // Agregar un nuevo elemento al carrito 
+        const elementoCarrito = document.createElement('li');
+        const precioTotal = (producto.precio * 1).toFixed(2); 
+        elementoCarrito.innerHTML = `
+            <div class="carrito-item">
+                <img src="${producto.imagen}" alt="${producto.nombre}" class="carrito-imagen">
+                <div class="carrito-info">
+                    <p>${producto.nombre} - $${producto.precio.toFixed(2)}</p>
+                    <button class="eliminar-producto" data-id="${producto.id}" onclick="eliminarProducto(event)">Eliminar</button>
+                    <span class="cantidad-producto">1</span>
+                    <span class="precio-total">${precioTotal}</span>
+                </div>
+            </div>
+        `;
+        elementoCarrito.querySelector('.carrito-imagen').style.width = '70px'; 
+        elementoCarrito.querySelector('.carrito-imagen').style.borderRadius = '50%';
+        listaCarrito.appendChild(elementoCarrito);
+    }
+
+    calcularTotal();
 }
 
 
-// Función para eliminar un producto del carrito
+document.addEventListener("DOMContentLoaded", async () => {
+    await mostrarProductos();
+    cargarCarritoDesdeLocalStorage();
+});
+
+document.getElementById('carrito-icon').addEventListener('click', function () {
+    let carrito = document.getElementById('carrito');
+    carrito.style.display = (carrito.style.display === 'none' || carrito.style.display === '') ? 'block' : 'none';
+});
+
 function eliminarProducto(e) {
     if (e.target.classList.contains('eliminar-producto')) {
         const productoId = e.target.getAttribute('data-id');
@@ -282,58 +366,41 @@ function eliminarProducto(e) {
     }
 }
 
-// Función para vaciar el carrito
 function vaciarCarrito() {
-    while (listaCarrito.firstChild) {
-        listaCarrito.removeChild(listaCarrito.firstChild);
-    }
-
+    listaCarrito.innerHTML = '';
     carritoCantidad = 0;
     actualizarIconoCarrito();
     calcularTotal();
 }
 
-// Función para calcular el total del carrito
-function calcularTotal() {
-    let total = 0;
-    const productosEnCarrito = document.querySelectorAll('#lista-carrito li');
-
-    productosEnCarrito.forEach((producto) => {
-        const precio = parseFloat(producto.textContent.split('-')[1].trim().substring(1));
-        total += precio;
-    });
-
-    totalCarrito.textContent = total.toFixed(2);
+function formatearNumero(numero) {
+    return new Intl.NumberFormat('es-AR', { minimumFractionDigits: 0, maximumFractionDigits: 3 }).format(numero);
 }
 
-// Función para actualizar el icono del carrito
+function calcularTotal() {
+    let total = 0;
+    const productosEnCarrito = document.querySelectorAll('#lista-carrito .carrito-item');
+
+    productosEnCarrito.forEach((producto) => {
+        const cantidad = parseInt(producto.querySelector('.cantidad-producto').textContent, 10);
+        const precioTotal = parseFloat(producto.querySelector('.precio-total').textContent);
+        total += precioTotal;
+    });
+
+    totalCarrito.textContent = formatearNumero(total);
+}
+
+
+
 function actualizarIconoCarrito() {
     carritoCantidadElement.textContent = carritoCantidad;
 }
 
-// Event listeners para el carrito
 listaCarrito.addEventListener('click', eliminarProducto);
 vaciarCarritoBtn.addEventListener('click', vaciarCarrito);
-
-// Script para mostrar/ocultar el carrito al hacer clic en el icono NUEVOOOOOOO
-document.getElementById('carrito-icon').addEventListener('click', function () {
-    var carrito = document.getElementById('carrito');
-    carrito.style.display = (carrito.style.display === 'none' || carrito.style.display === '') ? 'block' : 'none';
-});
-
-// Llamar a la función para mostrar todos los productos al cargar la página
-document.addEventListener("DOMContentLoaded", () => mostrarProductos());
-
-// Llamar a la función para cargar el carrito desde localStorage al cargar la página
-document.addEventListener("DOMContentLoaded", cargarCarritoDesdeLocalStorage);
-
-
-// Agregar evento de cambio al campo de búsqueda
 searchInput.addEventListener('input', () => {
     const terminoBusqueda = searchInput.value;
     mostrarProductos('', terminoBusqueda);
 });
 
-// Llamar a la función para mostrar todos los productos al cargar la página
-document.addEventListener("DOMContentLoaded", () => mostrarProductos());
